version3/requirements.txt
datasets
torch
pytorch-lightning
transformers
Pillow
PyWavelets
ptwt
torchvision
sentencepiece
protobuf
accelerate